Outsourced new generic functions
- Temporarily they are located in the app/ directory.
- Maybe we find better places
